CU-86953m5xr reworked order sync in a way that it will sync shipment only when order is fully shipped
