+++ Model/Carrier/Storekeeper.php
@@ -122,10 +122,10 @@ class Storekeeper extends AbstractCarrier implements CarrierInterface
     /**
      * @param int $storeId
      * @param array $apiRate
-     * @param float $orderTotal
+     * @param $orderTotal
      * @return int|mixed
      */
-    private function getShippingPrice(int $storeId, array $apiRate, float $orderTotal)
+    private function getShippingPrice(int $storeId, array $apiRate, $orderTotal)
     {
         if (array_key_exists('free_from_value_wt', $apiRate['shipping_method_price_flat_strategy'])
             && $orderTotal >= $apiRate['shipping_method_price_flat_strategy']['free_from_value_wt']) {
